+++ src/pages/Results/Results.module.css
@@ -5,6 +5,7 @@
   flex: 1;
   background-color: white;
   padding: 24px 0;
+  gap: 16px;
 }
 
 .center {
@@ -28,73 +29,4 @@
   padding: 8px 16px;
   font-size: 14px;
   cursor: pointer;
-}
-
-.channel_box {
-  display: flex;
-  width: 100%;
-}
-
-.channel_img_box {
-  flex: 3;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.channel_img {
-  width: 136px;
-  height: 136px;
-  aspect-ratio: 1;
-  overflow: hidden;
-  border-radius: 50%;
-}
-
-.channel_info {
-  flex: 4;
-  display: flex;
-  align-items: center;
-}
-
-.info_text {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.subs_btn {
-  border-radius: 18px;
-  font-size: 14px;
-  border: none;
-  color: white;
-  background-color: black;
-  padding: 8px 16px;
-  height: fit-content;
-  cursor: pointer;
-}
-
-.channel_detail {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-
-  font-size: 12px;
-  color: rgb(96, 96, 96);
-  padding-right: 24px;
-}
-
-.channel_stats {
-  display: flex;
-  gap: 4px;
-  align-items: center;
-}
-
-.channel_des {
-  flex: 1;
-  display: -webkit-box;
-  line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  text-overflow: ellipsis;
 }
