+++ src/pages/Results/ShortVideoItem.jsx
@@ -1,21 +1,15 @@
-import SkeletonImage from "../../components/SkeletonImage/SkeletonImage";
 import { formatISO } from "../../utils/date";
 import { formatHitCount } from "../../utils/hit";
-import styles from "./LongVideoItem.module.css";
+import styles from "./ShortVideoItem.module.css";
 
-const LongVideoItem = ({ item }) => {
+const ShortVideoItem = ({ item }) => {
   return (
     <div key={item.id.videoId} className={styles.video_box}>
       <div className={styles.video_img_box}>
-        <SkeletonImage
-          Image={
-            <img
-              src={item.snippet.thumbnails.high.url}
-              alt="video_thumbnail"
-              className={styles.video_img}
-            />
-          }
-          skeletonStyle={{ minHeight: "200px" }}
+        <img
+          src={item.snippet.thumbnails.high.url}
+          alt="video_thumbnail"
+          className={styles.video_img}
         />
       </div>
       <div className={styles.video_info}>
@@ -43,4 +37,4 @@ const LongVideoItem = ({ item }) => {
   );
 };
 
-export default LongVideoItem;
+export default ShortVideoItem;
