+++ src/pages/Results/ShortVideoItem.module.css
@@ -7,15 +7,26 @@
   flex: 3;
   padding-right: 10px;
   overflow: hidden;
-  min-height: 200px;
+  background: rgb(181, 182, 187);
+  background: radial-gradient(
+    circle,
+    rgba(181, 182, 187, 1) 0%,
+    rgba(105, 102, 102, 0.9864320728291317) 100%
+  );
+  margin-right: 10px;
+  border-radius: 12px;
+  overflow: hidden;
+
+  display: flex;
+  justify-content: center;
 }
 
 .video_img {
   overflow: hidden;
   border-radius: 8px;
   object-fit: cover;
-  width: 100%;
-  aspect-ratio: 580/280;
+  aspect-ratio: 9/16;
+  height: 100%;
 }
 
 .video_info {
